% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.R, R/choice_old.R
\name{choice}
\alias{choice}
\title{Alignment choice game to align internal reference planes \strong{(W)}.}
\usage{
choice(setup, savepaths, image_paths, touchup = NA, midpoint = FALSE,
  filetype = c("tif"), xpos = c(0, 660, 1250), brightness = 70,
  font_col = "white", font_size = 80, font_location = "+100+30",
  gravity = "southwest", choice_step = c(200, 100, 30, 10))

choice(setup, savepaths, image_paths, touchup = NA, midpoint = FALSE,
  filetype = c("tif"), xpos = c(0, 660, 1250), brightness = 70,
  font_col = "white", font_size = 80, font_location = "+100+30",
  gravity = "southwest", choice_step = c(200, 100, 30, 10))
}
\arguments{
\item{setup}{(required) Setup list is used to access internal reference AP coordinates.}

\item{savepaths}{(required) Paths to data directories.}

\item{image_paths}{(required) Paths to access registration images.}

\item{touchup}{(default = NA) Enter a vector of reference AP coordinates.
The choice game will replay for just those points.}

\item{midpoint}{(default = FALSE) Play the midpoint game following choice the
choice game.}

\item{filetype}{(optional, default = "tif") Image format to save.
Setting this argument as a vector of image formats, e.g. c("tif", "png")
will save both image types. Options: common image formats like "tiff", "png",
"jpeg", "gif", "rgb" or "rgba". See \link{magick} package documentation for more info
on the \code{\link[=image_write]{image_write()}} function}

\item{xpos}{(optional, default = c(0, 660, 1250)) A vector of x-positions in
pixels justified to the left of the main computer screen for the three choice windows.}

\item{brightness}{(optional, default = 70) Popup image brightness.}

\item{font_col}{(optional, default = "white") Image annotation font color.}

\item{font_size}{(optional, default = 80) Image annotation font size.}

\item{font_location}{(optional, default  = "+100+30")
Text position from center of gravity parameter in pixels}

\item{choice_step}{(optional, default= c(200,100,30,10))
Determines the successive "zoom" of images picked on either side of the center image.}

\item{setup}{(required) Setup list is used to access internal reference AP coordinates.}

\item{savepaths}{(required) Paths to data directories.}

\item{image_paths}{(required) Paths to access registration images.}

\item{touchup}{(default = NA) Enter a vector of reference AP coordinates.
The choice game will replay for just those points.}

\item{midpoint}{(default = FALSE) Play the midpoint game following choice the
choice game.}

\item{filetype}{(optional, default = "tif") Image format to save.
Setting this argument as a vector of image formats, e.g. c("tif", "png")
will save both image types. Options: common image formats like "tiff", "png",
"jpeg", "gif", "rgb" or "rgba". See \link{magick} package documentation for more info
on the \code{\link[=image_write]{image_write()}} function}

\item{xpos}{(optional, default = c(0, 660, 1250)) A vector of x-positions in
pixels justified to the left of the main computer screen for the three choice windows.}

\item{brightness}{(optional, default = 70) Popup image brightness.}

\item{font_col}{(optional, default = "white") Image annotation font color.}

\item{font_size}{(optional, default = 80) Image annotation font size.}

\item{font_location}{(optional, default  = "+100+30")
Text position from center of gravity parameter in pixels}

\item{choice_step}{(optional, default= c(200,100,30,10))
Determines the successive "zoom" of images picked on either side of the center image.}
}
\value{
The argument assigned to \emph{setup} is returned with aligned internal
z numbers matching internal AP coordinates.

The argument assigned to \emph{setup} is returned with aligned internal
z numbers matching internal AP coordinates.
}
\description{
User friendly way to play the wholebrain choice game to align internal
reference atlas plates. Automatically saves images of chosen aligned images in the
folder specified by \emph{savepaths$out_reference_aligned}. For use with mapping whole brain only.

User friendly way to play the wholebrain choice game to align internal
reference atlas plates. Automatically saves images of chosen aligned images in the
folder specified by \emph{savepaths$out_reference_aligned}. For use with mapping whole brain only.
}
