% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comment_overall_keys}
\alias{comment_overall_keys}
\title{Generate translation keys for overall comments}
\usage{
comment_overall_keys(ap, ar)
}
\arguments{
\item{ap}{Average difficulty (p-value).}

\item{ar}{Average discrimination (r-value).}
}
\value{
A character vector of translation keys.
}
\description{
Creates a vector of translation keys based on average
difficulty (ap) and discrimination (ar).
}
\examples{
comment_overall_keys(0.6, 0.35) # Medium, Keep
comment_overall_keys(0.2, 0.15) # Hard, Remove
}
