% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_similarity_vec}
\alias{cosine_similarity_vec}
\title{Calculate the cosine similarity between two vectors}
\usage{
cosine_similarity_vec(vec_1, vec_2)
}
\arguments{
\item{vec_1}{a numeric vector representing the first vector.}

\item{vec_2}{a numeric vector representing the second vector.}
}
\value{
a numeric value representing the cosine similarity between the two vectors.
}
\description{
Calculate the cosine similarity between two vectors
}
