% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Automatically generate ready-to-publish descriptive tables}
\usage{
create_table(
  df,
  stratify_by = NULL,
  var_removed = NULL,
  var_removed_strat = NULL,
  by_row = F
)
}
\arguments{
\item{df}{Data frame to be summarized as a table.}

\item{stratify_by}{Variable to create stratified tables from.}

\item{var_removed}{Variables to be removed from the overall summary.}

\item{var_removed_strat}{Variables to be removed from the stratified summary.}

\item{by_row}{Logical. If TRUE it will display the \% for each stratified group by row. By default, the \% by column is displayed.}
}
\value{
A list containing at most two data frames: one for the overall summary table, another for the stratified summary table.
}
\description{
This function loads a data frame and returns a statistical summary table
for each categorical or numeric variables. Optionally, it generates stratified
summary tables. Output tables can be customized by removing those
variables that are not of interest for either the overall summary or the stratified one.
}
