% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d_mode}
\alias{d_mode}
\title{Calculate the mode}
\usage{
d_mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode of the vector. Returns NA if the vector is empty.
}
\description{
Finds the most frequent value (the mode) in a vector.
}
\examples{
d_mode(c(1, 2, 2, 3, 3, 3, 4))
d_mode(c("a", "b", "a"))
}
