% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_id_cols}
\alias{detect_id_cols}
\title{Detect ID columns using regex}
\usage{
detect_id_cols(cols)
}
\arguments{
\item{cols}{A character vector of column names.}
}
\value{
A character vector of names that matched the ID pattern.
}
\description{
Searches a vector of column names for common ID-related patterns.
}
\examples{
detect_id_cols(c("Ad", "Soyad", "ogrenci no", "Madde1", "StudentID"))
}
