% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ensure_library}
\alias{ensure_library}
\title{This functions allow the user to choose a directory interactively
it is supposed to be cross-platform
comes from https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r}
\usage{
ensure_library(lib.name)
}
\description{
This functions allow the user to choose a directory interactively
it is supposed to be cross-platform
comes from https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r
}
