% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_savepaths.R
\name{ez_savepaths}
\alias{ez_savepaths}
\title{Automatically generates savepaths and save directories.}
\usage{
ez_savepaths(setup)
}
\arguments{
\item{setup}{(required) The output folder from the setup list is used to generate save directories}
}
\value{
\emph{savepaths} is a list containing savepaths to the data directories where analysis output will be stored.
}
\description{
User friendly way to create standardized save directories based on the setup information from the \emph{wb_setup} function.
Returns a list of savepaths to subfolders in the output folder.
}
\details{
Data directories are automatically created in the output folder.
}
