% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_sunburst.R
\name{ez_sunburst}
\alias{ez_sunburst}
\title{Generate a sunburst plot using a forward warped dataset.}
\usage{
ez_sunburst(dataset, rois = c("CH", "BS"), parent = TRUE)
}
\arguments{
\item{dataset}{(required) Whole brain dataset returned as an output of the
\code{\link[=forward_warp]{forward_warp()}} function.}

\item{rois}{(optional, default = c("CH", "BS")) The acronyms of the regions
of interest. Acronyms must follow  Allen Brain Atlas abbreviation
standards.}

\item{parent}{(optional, default = TRUE) TRUE allows multiple rois to be
plotted on the same sunburst, with the base parent layer set to 'grey'. If
FALSE, the base will be set to the first ROI. Other ROIS will be ignored
because the sunburst cannot accomodate more than one base layer.}
}
\value{
Returns an object of class sunburst.
}
\description{
Generate a sunburst plot using a forward warped dataset.
}
