% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezBrain.R
\docType{package}
\name{ezBrain}
\alias{ezBrain}
\alias{ezBrain-package}
\title{ezBrain: This package builds a pipeline interfacing with the \link{wholebrain}
package to process whole brain imaging datasets.}
\description{
The pipeline is split into five different parts, with the purpose
of each function per part explained below. Note that all functions that are
exclusively meant to be used with whole brain datasets will be marked with a
\strong{(W)} next to the function. Unmarked functions can be used on single slice
datasets across the brain.
}
\details{
For convention, if a return value is given by a function,
the recommended variable name is indicated in italics in the return
section of each function's help page. If this return value is a
required argument for another function in the pipeline, the
the recommended variable name will also be the same name as the argument.
}
\section{Part 1. Setup pipeline}{

\enumerate{
\item \code{\link[=setup_pl]{setup_pl()}} User friendly way to setup parameters for whole or
patrial brain pipeline analysis.
\item \code{\link[=im_sort]{im_sort()}}  A function to sort image paths for imaging datasets.
\item \code{\link[=ez_savepaths]{ez_savepaths()}} Generate savepaths and save directories.
}
}

\section{Part 2. Alignment (whole brain dataset)}{

\enumerate{
\item \code{\link[=choice]{choice()}} \strong{(W)} User friendly choice game to align internal
reference atlas plates.
\item \code{\link[=brainmorph]{brainmorph()}} \strong{(W)} Generate a brain morph plot.
\item \code{\link[=ez_interpolate]{ez_interpolate()}} \strong{(W)} Interpolate all AP and z numbers
for atlas plates in a whole brain project
}
}

\section{Part 3. Registration}{

\enumerate{
\item \code{\link[=ez_regi]{ez_regi()}} Console user interface built on top of
registration() function from the wholebrain package.
\item \code{\link[=regi_loop]{regi_loop()}} Automatically loops through the image registration
processs for the imaging dataset.
}
}

\section{Part 4. Segmentation and forward warping}{

\enumerate{
\item \code{\link[=filter_loop]{filter_loop()}} Loops through reference slices and allows user to
check/change filter settings for segmentation or registration throughout the
brain.
\item \code{\link[=seg_loop]{seg_loop()}} Loop through and segment images in the segmentation channel.
\item \code{\link[=clean_duplicates]{clean_duplicates()}} \strong{(W)} Duplicate cell count clean up of segmentation
output
\item \code{\link[=forward_warp]{forward_warp()}} Perform forward warp loop back onto atlas space using
segmentation and registration data.
}
}

\section{Part 5. Dataset manipulation and plotting}{

\enumerate{
\item \code{\link[=get_rois]{get_rois()}} Get subset of the forward warped dataframe of just regions of
interest.
\item \code{\link[=ez_sunburst]{ez_sunburst()}} Generate a sunburst plot using a forward warped dataset.
\item \code{\link[=get_tree]{get_tree()}} Create a dataframe of hierarchical region cell count data.
}
}

