% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_loop.R
\name{filter_loop}
\alias{filter_loop}
\title{Loop through reference slices}
\usage{
filter_loop(setup, channel = c("regi", "seg"), filter = NULL, console = TRUE)
}
\arguments{
\item{setup}{(required) Setup list from \code{\link[=setup_pl]{setup_pl()}}.}

\item{channel}{(optional, default = "regi") Which channel you are checking the filter for.}

\item{filter}{(optional, default = NULL) User entered custom filter.}

\item{console}{(optional, default = TRUE) Display console interface to modify filter parameters
instead of using the built-in GUI.}
}
\value{
Returns \emph{filter} with adjusted values.
}
\description{
This function will take a filter and loop through all images in a partial brain
or just the reference images across the brain in either the registration or segmentation channel.
The user can check qualitatively whether the filter needs adjustment after every image check. There is a console
menu option for the user to adjust filter parameters if they need. If the filter has been changed by the
user, either through the GUI or through the console menu, the function can reloop through
the start of the image sequence.
}
\seealso{
See also the \code{\link[wholebrain:segment]{wholebrain::segment()}} function in wholebrain.
}
