% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_warp.R
\name{forward_warp}
\alias{forward_warp}
\title{Forward warp back to atlas space}
\usage{
forward_warp(
  setup,
  segs,
  regis,
  filetype = c("tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp", "ps", "eps",
    "pdf"),
  savewarp = TRUE,
  saveschematic = TRUE,
  plane = "coronal",
  title = FALSE,
  mm.grid = TRUE,
  dev.size = c(5.4, 4.465),
  pch = 21,
  cex = 0.5,
  col = "black",
  scale.bar = FALSE,
  region.colors = TRUE
)
}
\arguments{
\item{setup}{(required) Setup list from \code{\link[=setup_pl]{setup_pl()}}.}

\item{segs}{(required) Segmentation data.}

\item{regis}{(required) Registration data.}

\item{filetype}{(optional, default = "tif") Image type to save as. See the
\emph{type} argument in the function \code{\link[=savePlot]{savePlot()}}. Options: "tif", "tiff",
"wmf", "emf", "png", "jpg", "jpeg", "bmp","ps", "eps", "pdf"}

\item{savewarp}{(optional, default = TRUE) Save forward warp plot image in
folder 'setup$savepaths$out_segmentation_warps'.}

\item{saveschematic}{(optional, default = TRUE) Save schematic plot images in
folder 'setup$savepaths$out_segmentation_schem'.}

\item{plane}{(optional, default = "coronal") Atlas plane to register to.
options: "coronal", "sagittal"}

\item{title}{(optional, default = FALSE) Title for the schematic plot.}

\item{mm.grid}{(optional, default = TRUE) Plot grid in schematic plot.}

\item{dev.size}{(optional, default = c(5.4, 4.465)). See same argument from
\code{\link[wholebrain:schematic.plot]{wholebrain::schematic.plot()}}.}

\item{pch}{(optional, default = 21) Graphical parameter for point shape.}

\item{cex}{(optional, default = 0.5) Graphical parameter for text size.}

\item{col}{(optional, default = "black") Graphical parameter for color.}

\item{scale.bar}{(optional, default = TRUE)  Display a measure bar in the
schematic plot}

\item{region.colors}{(optional, default = TRUE)  Display Allen Brain Atlas
colors in schemetic plot.}
}
\value{
Returns \emph{dataset} a variable storing all mapped segmentation data to
registration data.
}
\description{
Loop through all segmented cells and perform forward warp loop
back onto atlas space using registration and segmentation data. Options for
plotting a schematic plot and saving forward warp images. For a whole
brain, this function should be used after the segmentation list \emph{segs} has
been cleaned of duplicate cell counts. Note: if the imaging dataset is
large this will be a time intensive process. Processing time will be
printed once the function is finished.
}
\seealso{
See also \code{\link[wholebrain:schematic.plot]{wholebrain::schematic.plot()}}.
}
