% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_est.R
\name{get_est}
\alias{get_est}
\title{Get the estimates from the SMART-covariate model}
\usage{
get_est(model, stData, covarDat, covar)
}
\arguments{
\item{model}{the model output from SMART_covariate().}

\item{stData}{the stData used in SMART_covariate().}

\item{covarDat}{the covarDat used in SMART_covariate().}

\item{covar}{a character string selecting the covariate estimate you would like to get (one covariate at a time).}
}
\value{
a cell type proportion matrix

a cell type-specific gene expression matrix
}
\description{
Get the estimates from the SMART-covariate model
}
\examples{
\dontrun{
library(SMART)
data(MPOA)
mod <- SMART_covariate(stData=MPOA$stMat, markerGs=MPOA$markers,
                       covarDat=MPOA$covDat,covars='sex',
                       noMarkerCts=1,outDir='SMART_results', seed=1)
res <- get_est(model=mod$model, stData=MPOA$stMat,
               covarDat=MPOA$covDat, covar='sex')
}
}
