% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_itemexam_quant}
\alias{get_itemexam_quant}
\title{Safely get quantile from 'psychometric' package}
\usage{
get_itemexam_quant()
}
\value{
A numeric value for the quantile (default 0.27).
}
\description{
Tries to get the default \code{quant} argument from \code{psychometric::item.exam}.
Returns 0.27 if the package is not installed or an error occurs.
}
\examples{
get_itemexam_quant()
}
