% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_savepaths.R
\name{get_savepaths}
\alias{get_savepaths}
\title{Automatically generates savepaths and save directories.}
\usage{
get_savepaths(setup)
}
\arguments{
\item{setup}{(required) The output folder from the setup list is used to
generate save directories}
}
\value{
\emph{setup} is returned with \emph{setup$savepaths} filled containing
savepaths to the data directories where analysis output will be stored.
}
\description{
Create standardized save directories based on the setup
information from \code{\link[=setup_pl]{setup_pl()}}.
Returns a list of savepaths to subfolders in the output folder.
}
\details{
Data directories are automatically created in the output folder.
}
