% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Create a data frame of left/right counts and their ratio of regions of
interest}
\usage{
get_table(dataset, rois = c("CH", "BS"), base = "grey")
}
\arguments{
\item{dataset}{(required) Whole brain dataset returned as an output of the
\code{\link[=forward_warp]{forward_warp()}} function.}

\item{rois}{(optional, default = c("CH", "BS")) The acronyms of the regions
of interest. Acronyms must follow  Allen Brain Atlas abbreviation
standards.}

\item{base}{(optional, default = "grey") The base parent region you want
subregions traced back to. Note, base region MUST be a higher parent level
or the same level of the roi/rois provided to the \emph{rois} argument.}
}
\value{
Returns \emph{table}, a dataframe showing total, left and right cell
counts and count percentages according region acryonym.
}
\description{
Generates a dataframe showing region acronyms, their full name,
hierachical paths, total cell counts, left and right counts, and cell count
percentages.
}
\examples{
# Get cell counts of the hypothalamus
table <- get_table(dataset, rois = "HY", base = "HY")

# Get cell counts of all grey matter
table <- get_table(dataset, rois = "grey", base = "grey")

# Get cell counts of hypothalamus and all parent structures of the hypothalamus up to gray matter
table <- get_table(dataset, rois = "HY", base = "grey")
}
