% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Create a dataframe of hierarchical region cell count data.}
\usage{
get_tree(dataset, rois = c("CH", "BS"), base = "grey", subtract = TRUE)
}
\arguments{
\item{dataset}{(required) Whole brain dataset returned as an output of the
\code{\link[=forward_warp]{forward_warp()}} function.}

\item{rois}{(optional, default = c("CH", "BS")) The acronyms of the regions
of interest. Acronyms must follow  Allen Brain Atlas abbreviation
standards.}

\item{base}{(optional, default = "grey") The base parent region you want
subregions traced back to. This will be stored in the first hierarchy level
(col 1) of the dataframe. Note, base region MUST be a lower level parent of
all rois provided to the \emph{rois} argument.}

\item{subtract}{(optional, default = TRUE) Subtract the cell counts of all
child regions from parent regions. All counts number then uniquely belong
to a single acronym.}
}
\value{
Returns \emph{tree}, a dataframe storing the hierarchical relationship
between subregions. Cell counts, region colors, and cell count density is
also stored. Descriptions of columns of the dataframe are below:
\itemize{
\item Col 1-10. Each higher level column to the right is a sublevel of parent
column to the left.
\item Col 11. Cell counts of regions
\item Col 12. ABA color
}
}
\description{
Generates a dataframe of showing nested relationship of brain
regions, their cell counts, and their cell count densities. Output can also
be used for simple plots showing nested heirarchical relationships, such as
treemaps. As a convention, higher level columns will indicate child regions
while lower
level columns indicate parent regions.
}
