% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_brain.R
\name{gg_brain}
\alias{gg_brain}
\title{Function to plot planes in `ggplot2` environment}
\usage{
gg_brain(contour_list, grayscale = TRUE, color = FALSE, fill = FALSE)
}
\arguments{
\item{contour_list}{comes from `prep_data(regi)` function}
}
\description{
The idea behind this script is to make something like wholebrain::plot.outlines(regi, TRUE) but in ggplot2
}
