% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassbrain2.R
\name{glassbrain2}
\alias{glassbrain2}
\title{Modified glassbrain function}
\usage{
glassbrain2(
  dataset,
  high.res = "OFF",
  dim = c(720, 1080),
  device = TRUE,
  col = "region",
  cex = 0.5,
  hemisphere = "right",
  spheres = FALSE,
  alpha = 1,
  laterality = TRUE,
  plane = "coronal",
  jitter = FALSE
)
}
\arguments{
\item{dataset}{(required) Cell count dataset}

\item{high.res}{(default = "OFF") TRUE or FALSE sets high or low resolution glassbrain. "OFF"
doesn't plot plot the glassbrain. Options: "OFF", TRUE, FALSE}

\item{dim}{(default = c(720, 1080))  Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{device}{(default = TRUE) Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{col}{(default = "region") Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{cex}{(default = 0.5) Graphical parameter for text size.}

\item{hemisphere}{(default = "right") Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{spheres}{(default = FALSE) Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{alpha}{(default = 1) Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{laterality}{(default = TRUE) Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{plane}{(default = "coronal") Original argument from \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}}.}

\item{jitter}{(default = FALSE) If TRUE, add jitter in z-axis to fill space between z-planes.}
}
\description{
A modified version of \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}} that
\enumerate{
\item Gives the user an option to plot the "jitter" in cell placement in original glassbrain function
(the jitter gives a more space filled look) .
\item Does not show the glass brain display when high.res is set to "OFF" (otherwise set to TRUE or FALSE).
}
}
\seealso{
See also the function \code{\link[wholebrain:glassbrain]{wholebrain::glassbrain()}} from the wholebrain package.
}
