% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_AP_match.R
\name{inspect_AP_match}
\alias{inspect_AP_match}
\title{User friendly way to inspect whether `match_image_to_atlas` was good}
\usage{
inspect_AP_match(match_df)
}
\arguments{
\item{match_df}{dataframe with image_paths and assigned APs (output of `match_image_to_atlas`)}
}
\description{
Shows image and atlas to check whether tey match, let's user flag correct/incorrect.
}
