% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate all AP and z numbers for atlas plates in a whole brain project \strong{(W)}.}
\usage{
interpolate(setup, method = "l")
}
\arguments{
\item{setup}{(required) Setup list will be used as the return variable.}

\item{method}{(optional, default = "l") Interpolation method.
Options: "l" (linear interpolation), "s" (monotonic spline interpolation)}
}
\value{
The argument assigned to \emph{setup} is returned with
interpolated AP and z numbers based on the reference AP and z planes.
}
\description{
This function interpolates all corresponding z and AP values for atlas plates that are not reference plates.
For use with mapping whole brain only.
}
