% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_wb.R
\name{interpolate_wb}
\alias{interpolate_wb}
\title{Interpolate all AP and z numbers for atlas plates in a whole brain project.}
\usage{
interpolate_wb(setup, method = "l")
}
\arguments{
\item{setup}{(required) Setup list will be used as the return variable.}

\item{method}{(optional, default = "l") Interpolation method.
Options: "l" (linear interpolation), "s" (monotonic spline interpolation)}
}
\description{
This function interpolates all corresponding z and AP values for atlas plates that are not reference plates.
For use with mapping whole brain only.
}
