% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_scored_01}
\alias{is_scored_01}
\title{Check if a vector is scored 0/1}
\usage{
is_scored_01(vec)
}
\arguments{
\item{vec}{The vector to check.}
}
\value{
TRUE if the vector is 0/1 scored, FALSE otherwise.
}
\description{
Detects if a vector (after removing NAs) contains only 0 and 1.
}
\examples{
is_scored_01(c(1, 0, 1, 0, NA))
is_scored_01(c(1, 0, 2, 0))
is_scored_01(c("A", "B", "C"))
}
