% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltable.R
\name{modeltable}
\alias{modeltable}
\title{Generate summary tables from INLA, GLM or HR model outputs}
\usage{
modeltable(result, model)
}
\arguments{
\item{result}{Output from the model function}

\item{model}{Either "INLA", "GLM" of "HR", depending on the model used.}
}
\value{
A data frame summarizing the main results from the model.
}
\description{
This function loads the resulting output of either INLA, GLM or HR models,
and returns a ready-to-publish summary table from it.
}
