% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_letter}
\alias{norm_letter}
\title{Normalize letter grades}
\usage{
norm_letter(x)
}
\arguments{
\item{x}{A vector, typically character.}
}
\value{
A character vector of normalized grades (A-E) or NA.
}
\description{
Cleans and validates a vector of characters, keeping only
standard letter grades (A, B, C, D, E).
}
\examples{
norm_letter(c(" a ", "B", "c", "F", "d", NA))
}
