% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_lc_bin}
\alias{parse_lc_bin}
\title{Parse 1/0 coded data}
\usage{
parse_lc_bin(x)
}
\arguments{
\item{x}{A vector of potential 1/0 scores.}
}
\value{
An integer vector (1, 0, or NA).
}
\description{
Validates that a vector contains only 1s, 0s, or NAs.
}
\examples{
parse_lc_bin(c(1, 0, "1", "0", 2, "A", NA))
}
