% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_lc_raw}
\alias{parse_lc_raw}
\title{Parse raw continuous/Likert scores}
\usage{
parse_lc_raw(x)
}
\arguments{
\item{x}{A vector (typically character) of raw scores.}
}
\value{
A numeric vector of cleaned scores.
}
\description{
Cleans a vector of potential scores, converting to numeric
and removing invalid or out-of-range values.
}
\examples{
parse_lc_raw(c("10", "5.5", "0", "-2", "ikiyüz", NA, "1000001"))
}
