\name{plot.gnsccv}
\alias{plot.gnsccv}
\title{
Plot function for S3 class "gnsccv"
}
\description{
Plot sparsity level information and heatmap from the gnsc.cv
}
\usage{
\method{plot}{gnsccv}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"gnsccv"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\seealso{
\code{\link{gnsc.cv}}
}