\name{plot.TCE}
\alias{plot.TCE}
\title{
Plot function for S3 class "TCE"
}
\description{
Plot sparsity level information and 3 typical sparse graphs from the correlation graph path
}
\usage{
\method{plot}{TCE}(x, align = FALSE, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"TCE"} 
}
  \item{align}{
If \code{align = FALSE}, 3 plotted graphs are aligned
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\seealso{
\code{\link{TCE}}
}