% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Function to plot planes in `ggplot2` environment}
\usage{
prep_data(regi, ordered_filter_list)
}
\arguments{
\item{regi}{object from any `registration` function(s)}

\item{ordered_filtered_list}{a list of the filters feed to
the regi object in the same order (check names(ordered_filtered_list) matches the regi[[i]])}
}
\description{
This function will make a table with the information we need to provide to the cropping function
}
