\name{print.gnsc}
\alias{print.gnsc}
\title{
Print function for S3 class "gnsc"
}
\description{
Print the information about threshold values, selected variables, fitted errors, predicted values
}
\usage{
\method{print}{gnsc}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"gnsc"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\seealso{
\code{\link{gnsc.train}}
}