\name{print.gnsccv}
\alias{print.gnsccv}
\title{
Print function for S3 class "gnsccv"
}
\description{
Print the information about threshold values, selected variables, cross-validation errors
}
\usage{
\method{print}{gnsccv}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"gnsccv"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\seealso{
\code{\link{gnsc.cv}}
}