\name{print.TCA}
\alias{print.TCA}
\title{
Print function for S3 class "TCA"
}
\description{
Print the information about model, algorithm and results
}
\usage{
\method{print}{TCA}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"TCA"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\seealso{
\code{\link{TCA}}
}