\name{print.TCE}
\alias{print.TCE}
\title{
Print function for S3 class "TCE"
}
\description{
Print the information about the model usage, the graph path length, graph dimension, sparsity level
}
\usage{
\method{print}{TCE}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"TCE"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\seealso{
\code{\link{TCE}}
}