% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_atlas.R
\name{pull_atlas}
\alias{pull_atlas}
\title{Pull up atlas plate in graphics device.}
\usage{
pull_atlas(AP, font_size = 400, font_location = "+250+150", xpos = 0)
}
\arguments{
\item{AP}{Anterior-posterior mouse atlas coordinate.}

\item{xpos}{x-position in pixels justified to the left of the main computer
screen}
}
\description{
Enter in an AP coordinate from the Allen Mouse Brain Atlas and
  pull up the atlas in a graphics device.
}
