% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{q_index}
\alias{q_index}
\title{Calculate q-index (1 - p)}
\usage{
q_index(p)
}
\arguments{
\item{p}{A numeric value or vector (item difficulty).}
}
\value{
A numeric value or vector (1 - p).
}
\description{
A simple helper to calculate the inverse of the p-value (difficulty index).
}
\examples{
q_index(0.8)
q_index(c(0.2, 0.5, 0.7))
}
