% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regi_loop.R
\name{regi_loop}
\alias{regi_loop}
\title{Automatically loops through the image registration process}
\usage{
regi_loop(
  setup,
  filter = NULL,
  regis = NULL,
  plane = "coronal",
  closewindow = TRUE,
  filetype = c("tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp", "ps", "eps",
    "pdf"),
  autoloop = FALSE,
  touchup = FALSE,
  reference = FALSE,
  popup = TRUE,
  brightness = 70,
  font_col = "white",
  font_size = 40,
  font_location = "+100+100",
  gravity = "southwest",
  width = 18,
  height = 10.2
)
}
\arguments{
\item{setup}{(required) Setup list from \code{\link[=setup_pl]{setup_pl()}}.}

\item{filter}{(required) Filter optimized for registration of images in the registration channel.}

\item{regis}{(optional, default = NULL) List for storing registration output.
On the first run of this function, the regis list will be automatically created in the global environment.
If it exists in the global environment, this
argument \strong{must} be set to \emph{regis = regis}.}

\item{plane}{(optional, default = "coronal") Atlas plane to register to. options: "coronal", "sagittal"}

\item{closewindow}{(optional, default = TRUE) If TRUE, windows will close after every correspondance point alteration.}

\item{filetype}{(optional, default = "tif") Image type to save as. See the \emph{type} argument in the function \code{\link[=savePlot]{savePlot()}}.
Options: "tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp","ps", "eps", "pdf"}

\item{autoloop}{(optional, default = FALSE) Run autoloop to save wholebrain's first pass at
registering all registration planes. This argument takes precendence over arguments 'touchup' and 'reference'}

\item{touchup}{(optional, default = FALSE) Run registration loop on a subset of registration plates.
Options:
\itemize{
\item touchup = TRUE runs a user friendly interface allowing user to enter which plates they want to fix.
\item touchup = numeric vector of plate numbers user wants to fix (plates should be whole integers).
}}

\item{reference}{(optional, default = FALSE)  Run registration loop on reference planes only if TRUE.
This argument takes precendence over argument 'touchup' UNLESS the project is not a whole brain project.
Entering argument TRUE on a partial brain project will be ignored.}

\item{popup}{(optional, default = TRUE) Popup the original raw image to be registered.
This can help with internal structure visualization.}

\item{brightness}{(optional, default = 70) Popup image brightness.}

\item{font_col}{(optional, default = "white")}

\item{font_size}{(optional, default = 40)}

\item{font_location}{(optional, default  = "+100+100") Move text from center of gravity parameter in pixels.}

\item{gravity}{(optional, "southwest") Location of text annotation. e.g. "southwest" means bottom left of screen.}

\item{width}{(default = 18) Registration plotting window width in inches.}

\item{height}{(default = 10.2 ) Registration plotting window height in inches.}
}
\description{
User friendly code to alter atlas to image registration.
When run with defaults, this function will automatically create a list called \emph{regis} with the length of setup$regi_z.
Each element of that list will be NULL until manual correction to that registration is done.
The function will loop through every single image listed in setup$regi_z in order.
A simple user interface built into the console is provided to improve each registration.
The user has the option to save the environment after every registration.
}
\details{
This functions creates a vector list of registration information for each registration slice.
Note: This is not a return. The vector \emph{regis} is automatically created.
}
