% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_table.R
\name{region_table}
\alias{region_table}
\title{Plot a region table.}
\usage{
region_table(dataset, savepaths, saveplot = TRUE,
  region.lab = "Input region:", xlab = "Cell count")
}
\description{
Plot a region table based on cell count density of the
data.frame information assigned to variable \link{dataset}.
}
