% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration2.R
\name{registration2}
\alias{registration2}
\title{Console user interface built on top of \code{\link[=registration]{registration()}} function from the wholebrain package.}
\usage{
registration2(
  input,
  coordinate = NULL,
  plane = "coronal",
  right.hemisphere = NULL,
  interpolation = "tps",
  intrp.param = NULL,
  brain.threshold = 200,
  blurring = c(4, 15),
  pixel.resolution = 0.64,
  resize = (1/8)/4,
  correspondance = NULL,
  resolutionLevel = c(4, 2),
  num.nested.objects = 0,
  plateimage = FALSE,
  forward.warp = FALSE,
  filter = NULL,
  output.folder = "../",
  batch.mode = FALSE,
  channel = 0,
  verbose = TRUE,
  closewindow = TRUE,
  width = 18,
  height = 10.2
)
}
\arguments{
\item{input}{(required) A character vector consisting of the full path name to 16-bit raw tif image files.}

\item{coordinate}{(default = NULL) Matching coordinates in the main plane to be registered to (in millimeters).}

\item{plane}{(default = "coronal") The main plane to register to: "coronal", "sagittal".}

\item{right.hemisphere}{(default = NULL)}

\item{interpolation}{(default = "tps")}

\item{intrp.param}{(default = NULL)}

\item{brain.threshold}{(default = 200) An integer value, which determinesthe segmentation of the brain slice.}

\item{blurring}{(default = c(4, 15))}

\item{pixel.resolution}{(default = 0.64)}

\item{resize}{(default = (1/8)/4)}

\item{correspondance}{(default = NULL)}

\item{resolutionLevel}{(default = c(4, 2))}

\item{num.nested.objects}{(default = 0)}

\item{plateimage}{(default = FALSE)}

\item{forward.warp}{(default = FALSE)}

\item{filter}{(default = NULL)}

\item{output.folder}{(default = "../")}

\item{batch.mode}{(default = FALSE)}

\item{channel}{(default = 0)}

\item{verbose}{(default = TRUE) If TRUE diagnostic output is written to the R console. Default is true.}

\item{closewindow}{(default = TRUE) If TRUE, windows will close after every correspondance point alteration.}

\item{width}{(default = 18) Plotting window width in inches.}

\item{height}{(default = 10.2 ) Plotting window height in inches.}
}
\value{
returns \emph{regi}, a vector list of registration information for each registration slice.
}
\description{
Allows you to alter correspondance points with user interface with the console
This uses the original \code{\link[=registration]{registration()}} function from wholebrain and retains the same arguments.
}
\seealso{
See also \code{\link[wholebrain:registration]{wholebrain::registration()}} for the original registration function from wholebrain.
}
