% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_AP.R
\name{rename_AP}
\alias{rename_AP}
\title{User friendly way to look for conflicts of alignment and rename images with Z planes}
\usage{
rename_AP(match_df)
}
\arguments{
\item{match_df}{df that comes from `match_image_to_atlas` function}
}
\description{
Receives df from previous pipeline step, looks for conflicts and renames the images adding Z numbers
}
