% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_AP_order.R
\name{rename_AP_order}
\alias{rename_AP_order}
\title{User friendly way to put images into order and rename them before setup}
\usage{
rename_AP_order(img_folder, filetype = c(".tif"))
}
\arguments{
\item{img_folder}{path of folder where images are stored}

\item{filetype}{type of the files, pattern fed to `list.files(...)`}
}
\description{
Asks user for folder where images are stored.
  If image titles are not in AP order, it prompts user to establish order
  and rename files properly into Z planes
}
