% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_loop.R
\name{seg_loop}
\alias{seg_loop}
\title{Loop through and segment images in the segmentation channel}
\usage{
seg_loop(
  setup,
  filter,
  numthresh = 8,
  downsample = 0.25,
  post = NULL,
  pre = NULL,
  get.contour = FALSE,
  channel = 0
)
}
\arguments{
\item{setup}{(required) Setup list from \code{\link[=setup_pl]{setup_pl()}}.}

\item{filter}{(required) Filter optimized for segmentation of images in the segmentation channel.}

\item{numthresh}{(optional, default = 8) Original argument from \code{\link[wholebrain:segment]{wholebrain::segment()}}.}

\item{downsample}{(optional, default = 0.25) Original argument from \code{\link[wholebrain:segment]{wholebrain::segment()}}.}

\item{post}{(optional, default = NULL) Original argument from \code{\link[wholebrain:segment]{wholebrain::segment()}}.}

\item{pre}{(optional, default = NULL) Original argument from \code{\link[wholebrain:segment]{wholebrain::segment()}}.}

\item{get.contour}{(optional, default = FALSE) Original argument from \code{\link[wholebrain:segment]{wholebrain::segment()}}.}

\item{channel}{(optional, default = 0) Original argument from \code{\link[wholebrain:segment]{wholebrain::segment()}}.}
}
\value{
returns \emph{segs} A list of length 2. Each list element is described below:
\enumerate{
\item A vector of the z-image number corresponding to each segmented slice stored in the 2nd element.
\item A list of segmentation information for each segmentated image.
}
}
\description{
Loop through the z images in the segmentation folder path. For a
whole brain, every N images is segmented, where N is equal to
\code{setup$seg_step.} For a partial brain, all the images in \code{setup$regi_z} are
segmented. The function will return a list called \emph{segs} of length 2. Note:
if the imaging dataset is large this will be a time intensive process.
Processing time will be printed once the function is finished.
}
\seealso{
See also the \code{\link[wholebrain:segment]{wholebrain::segment()}} function.
}
