% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_pl.R
\name{setup_pl}
\alias{setup_pl}
\title{User friendly way to setup whole brain pipeline analysis}
\usage{
setup_pl(setup = NULL)
}
\arguments{
\item{setup}{(optional) Allows the user to modify a setup list previously
generated by this function.}
}
\value{
\emph{setup} is a list of length 7 or length 16 with parameters for the pipeline analysis.
}
\description{
Asks user for setup information for wholebrain analysis.
Generates a setup list of length 9 if the user is registering a partial
brain or of length 18 if the user is mapping a whole brain. If a setup list
already exists, calling this function will simply ask the user to verify
setup info. Info can be changed if necessary. However, be conservative with
changing setup parameters, especially if different users work on the same brain.
}
\details{
Descriptions of list elements for a \strong{partial brain}.
\enumerate{
\item Animal ID
\item User's Initials.
\item Registration channel image folder path (normally an autofluorescence
channel)
\item Segmentation channel image folder path (can be different than the
Registration channel)
\item Output folder to store data
\item All AP coordinates ranging from the first to last AP. User manually
entered.
\item All z image numbers matching element 6. User manually entered.
\item Savepaths list containing paths to the data directories. Will be
NULL until \code{\link[=get_savepath]{get_savepath()}} is run.
\item List of sorted image paths. Will be NULL until \code{\link[=im_sort]{im_sort()}} is run.
}

Descriptions of list elements for a \strong{whole brain}.
\enumerate{
\item Animal ID
\item User's Initials
\item Registration channel image folder path (normally an autofluorescence
channel)
\item Segmentation channel image folder path (can be the same as the
Registration channel)
\item Output folder to store data
\item Spacing between adjacent z images (mm)
\item Spacing between registrations
(mm). Default is 0.100, meaning registrations are performed every 100 um.
Recommended input is are multiples of 0.100. Beware! Larger spacing between
registrations means less manual work but less accurate mapping in the rostral
caudal axis.
\item Integer step size between z-images used for segmentation. Default is 1 (1
skips no images, 2 skips every other, etc.).
\item First AP coordinate
\item First z image. Note: Keep all images of the stack together, even if the
first usable image isn't 1.
\item Last AP coordinate
\item Last z image
\item AP coordinates of internal reference planes. Default coordinates are
1.91, 1.10, -.42, -0.93 , -1.94 , -2.95, -3.96. They correspond to PFC, NAc,
antHyp, startHip, posHyp, VTA, PAG, respectively.
\item z image numbers matching AP coordinates (14). Will be NULL until
\code{\link[=choice]{choice()}} is run.
\item All AP coordinates ranging from the first to last AP.
Will be NULL until \code{\link[=interpolate]{interpolate()}} is run.
\item All z image numbers matching 16. Will be NULL until \code{\link[=interpolate]{interpolate()}} is
run.
\item Savepaths list containing paths to the data directories. Will be
NULL until \code{\link[=get_savepaths]{get_savepaths()}} is run.
\item List of sorted image paths. Will be NULL until \code{\link[=im_sort]{im_sort()}} is run.
}
}
