% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SMART.binary}
\alias{SMART.binary}
\title{Simulating a 2-Stages SMART with Binary Final Outcome}
\usage{
SMART.binary(n, p_A, p_AC, p_AD, p_AA, p_B, p_BE, p_BF, p_BB)
}
\arguments{
\item{n}{sample size of the simulated trial.}

\item{p_A, p_B}{response rates to initial treatments.}

\item{p_AC, p_AD, p_AA, p_BE, p_BF, p_BB}{response rates to stage-2 treatments for each sequence of treatments.}
}
\value{
a data frame with \code{n} rows and 4 columns. For each observation the
matrix provides the first treatment label, sequence of treatments label,
final outcome, and response indicator to the 1-stage intervention
(1 for responders and 0 for non-responders).
}
\description{
simulates a 2-stages SMART with binary final outcome in the
standard design where responders to the first treatment are not
re-randomized.
}
\examples{
SMART.binary(60, 0.5, 0.3, 0.3, 0.2, 0.5, 0.2, 0.2, 0.2 )
}
