% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SMART.continuous}
\alias{SMART.continuous}
\title{Simulating a 2-Stages SMART With Continuous Final Outcome}
\usage{
SMART.continuous(
  n,
  p_A,
  p_B,
  phi_1,
  phi_2,
  phi_3,
  phi_4,
  phi_5,
  phi_6,
  sd_aa,
  sd_ac,
  sd_ad,
  sd_bb,
  sd_be,
  sd_bf
)
}
\arguments{
\item{n}{sample size of the simulated trial.}

\item{p_A, p_B}{response rates to initial treatments.}

\item{phi_1, phi_2, phi_3, phi_4, phi_5, phi_6}{parameters of the outcome conditional mean model.}

\item{sd_aa, sd_ac, sd_ad, sd_bb, sd_be, sd_bf}{standard deviation values of the conditional
outcome distribution.}
}
\value{
A data frame with \code{n} rows and 4 columns. For each observation, the
matrix provides the first treatment label, sequence of treatments label,
final outcome, and response indicator to the 1-stage intervention
(1 for responders and 0 for non-responders).
}
\description{
simulates a 2-stages SMART with binary final outcome in the
standard design where responders to the first treatment are not
re-randomized.
}
\examples{
SMART.continuous(n = 100, p_A = 0.5, p_B = 0.5, phi_1 = 10, phi_2 = 5, phi_3 = -15,
phi_4 = -3, phi_5 = 10, phi_6 = -3, sd_aa = 2, sd_ac = 2, sd_ad = 2,
sd_bb = 2, sd_be = 3, sd_bf = 2)


}
