\name{smart-internal}
\title{Internal smart functions}
\alias{gnsc.heatmap}
\alias{gnsc.icov}
\alias{gnsc.predict}
\alias{gnsc.restruc}
\alias{minid}
\alias{smart.npn}
\description{Internal smart functions}
\author{
Fang Han, Han Liu \cr
Maintainer: Fang Han<fhan@jhsph.edu>
}
\details{
These are not intended for use by users. Please refer to \code{gnsc.train(), gnsc.cv(), TCA(), TCE()}. 
}
