% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SMART.power}
\alias{SMART.power}
\title{Computing the Power for a 2-Stages SMART}
\usage{
SMART.power(
  n,
  vn,
  sigma_n,
  theta_0,
  sigma_0,
  theta_d,
  sigma_d,
  epsilon,
  upper_int = Inf
)
}
\arguments{
\item{n}{sample size of the trial.}

\item{vn}{degrees of freedom of the inverse chi-squared distribution of tau_1+tau_2.}

\item{sigma_n}{non-centrality parameter of the inverse chi-squared distribution of tau_1+tau_2.}

\item{theta_0}{mean of the analysis prior Normal density.}

\item{sigma_0}{standard deviation of the analysis prior Normal density.}

\item{theta_d}{mean of the design prior Normal density.}

\item{sigma_d}{standard deviation of the design prior Normal density.}

\item{epsilon}{bayesian equivalent of type I error. 1-epsilon represents the
Bayesian significance level.}

\item{upper_int}{upper limit of the integral that marginalizes the Bayesian power function.
The default is \code{Inf}.}
}
\value{
Bayesian marginal power value.
}
\description{
Computes the Bayesian marginal power for the comparison of two strategies
that start with a different initial treatment in a 2-stages SMART in the
standard design where responders to the first treatment are not re-randomized.
}
