\docType{package}
\name{smart-package}
\alias{smart-package}
\alias{smart}

\title{Steve Martin's Assorted R Toolbox}

\description{
A collection of functions for R that I sometimes find useful.
}

\seealso{
\url{https://github.com/marberts/smart}
}

\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com}
}
