% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SMART.ss}
\alias{SMART.ss}
\title{Estimating the Sample Size of a 2-Stages SMART}
\usage{
SMART.ss(
  n_grid,
  theta_0,
  sigma_0,
  theta_d,
  sigma_d,
  power,
  epsilon,
  tr1 = NULL,
  tr2 = NULL,
  outcome = NULL,
  R = NULL,
  id1_study = NULL,
  id2_study = NULL,
  id1_ref = NULL,
  id2_ref = NULL,
  v_p = 0.5,
  k_p = 1,
  u_p = 0,
  sigma_p = 4,
  vn = NULL,
  sigma_n = NULL,
  alt = "one-sided",
  upper_int = Inf,
  save_grid = F
)
}
\arguments{
\item{n_grid}{vector of sample size values used to search for the optimal sample size.}

\item{theta_0}{mean of the analysis prior Normal density. If \code{theta_0="pilot"}, this density
is centered at the difference between mean strategies estimated from the pilot study.}

\item{sigma_0}{standard deviation of the analysis prior Normal density.}

\item{theta_d}{mean of the design prior Normal density.}

\item{sigma_d}{standard deviation of the design prior Normal density.}

\item{power}{target power level.}

\item{epsilon}{bayesian equivalent of type I error. 1-epsilon represents the
Bayesian significance level.}

\item{tr1}{vector of length \code{n} indicating the stage-1 interventions.}

\item{tr2}{vector of length \code{n} indicating the stage-2 interventions.}

\item{outcome}{vector of final outcomes.}

\item{R}{vector of responses to the stage-1 intervention
(1 for responders and 0 for non-responders).}

\item{id1_study}{label of the stage-1 intervention used in \code{tr1} of the strategy 1.}

\item{id2_study}{label of the stage-2 intervention used in \code{tr2} of the strategy 1.}

\item{id1_ref}{label of the stage-1 intervention used in \code{tr1} of the strategy 2.}

\item{id2_ref}{label of the stage-2 intervention used in \code{tr2} of the strategy 2.}

\item{v_p, k_p, u_p, sigma_p}{parameters of the Normal-inverse-chi-squared (NIX)
prior distribution used to estimate the posterior distribution of the variance
components. The default choice is \code{v_p=0.5,k_p=1,u_p=0,sigma_p=4}.}

\item{vn}{degrees of freedom of the inverse chi-squared distribution of tau_1+tau_2.}

\item{sigma_n}{non-centrality parameter of the inverse chi-squared distribution of tau_1+tau_2.}

\item{alt}{Alternative hypothesis type. The default option is "one-sided". Use "two-sided" for
a two-sided alternative hypothesis type.}

\item{upper_int}{upper limit of the integral that marginalizes the Bayesian power function.
The default is \code{Inf}. If an error occurs regarding the integral being divergent
or the upper bound of n_grid being high enough despite increasing it, try selecting
a large finite value for \code{upper_int}.}

\item{save_grid}{logical; if TRUE, a the estimated power for each value of \code{n_grid}
is stored. The default is FALSE.}
}
\value{
By default, the function returns two values: the
estimated sample size and corresponding power. If \code{save_grid = F},
an additional data frame containing the estimated power for each value of
\code{n_grid} is returned.
}
\description{
Computes the required sample size
in a 2-stages SMART to achieve the desired power level in the
standard design where responders to the first treatment are not re-randomized. To estimate
the variance components, the function needs either data from the pilot study through
the parameters \code{(tr1, tr2, outcome, R, id1_study, id2_study, id1_ref, id2_ref)} or
the direct specification of \code{vn} and \code{sigma_n}.
}
\examples{
set.seed(123)
p_A <- 0.5 ; p_B <- 0.5
phi_1 <- 10 ; phi_2 <- 5 ; phi_3 <- -15; phi_4 <- -3; phi_5 <- 10; phi_6 <- -3
sd_aa = 2 ; sd_ac = 2 ; sd_ad = 2 ; sd_bb = 2 ; sd_be = 3; sd_bf = 2
sim <- SMART.continuous(n = 100, p_A = p_A, p_B = p_B, phi_1 = phi_1, phi_2 =phi_2,
                          phi_3=phi_3, phi_4=phi_4,phi_5=phi_5,  phi_6=phi_6,
                          sd_aa = sd_aa, sd_ac = sd_ac,
                          sd_ad = sd_ad,sd_bb = sd_bb,
                          sd_be = sd_be,sd_bf = sd_bf)

 SMART.ss(n_grid = seq(1,7000), theta_0=0, sigma_0 = 0.2,
          theta_d=2, sigma_d = 0.1, power = 0.9, epsilon = 0.05,
          tr1 = sim$tr1, tr2 = sim$tr2, outcome = sim$outcome, R = sim$R,
          id1_study = "A", id2_study = "AB", id1_ref = "B", id2_ref = "BE")

}
