% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART_2S.R
\name{SMART_2S}
\alias{SMART_2S}
\title{Spatial transcriptomics deconvolution with the base model of SMART}
\usage{
SMART_2S(
  stData,
  markerGs_S1,
  markerGs_S2,
  noMarkerCts_S1 = 1,
  noMarkerCts_S2 = 1,
  CT_OI,
  CT_similar = NULL,
  outDir = "SMART_results",
  seed = 1,
  iterations = 2000,
  priors_S1 = NULL,
  priors_S2 = NULL
)
}
\arguments{
\item{stData}{a spot(rows)-by-gene(columns) spatial transcriptomics matrix.}

\item{markerGs_S1}{a list with marker genes for major cell types.}

\item{markerGs_S2}{a list with marker genes for subtypes of the cell type of interest.}

\item{noMarkerCts_S1}{the number of cell types without marker genes in stage one.}

\item{noMarkerCts_S2}{the number of cell types without marker genes in stage two.}

\item{CT_OI}{cell type of interest}

\item{CT_similar}{a cell type that is transcriptomically similar to the cell type of interest (can be a character string indicating a cell type name, 'auto' for automatic selection or NULL)}

\item{outDir}{the path to the output directory.}

\item{seed}{the starting value.}

\item{iterations}{the number of iterations (default: 2000).}

\item{priors_S1}{do not change the priors if you don't know what they are.}

\item{priors_S2}{do not change the priors if you don't know what they are.}
}
\value{
a model object

a model

a cell type proportion matrix
}
\description{
Spatial transcriptomics deconvolution with the base model of SMART
}
\examples{
\dontrun{
library(SMART)
data(NSCLC)
res <- SMART_2S(NSCLC$stData, NSCLC$markers_S1, NSCLC$markers_S2,
                noMarkerCts_S1=1, noMarkerCts_S2=1,
                CT_OI='DC',CT_similar='macrophage',
                outDir='SMART_results')
}
}
