% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART_base.R
\name{SMART_base}
\alias{SMART_base}
\title{Spatial transcriptomics deconvolution with the base model of SMART}
\usage{
SMART_base(
  stData,
  markerGs,
  noMarkerCts = 1,
  outDir = "SMART_results",
  seed = 1,
  iterations = 2000,
  priors = NULL
)
}
\arguments{
\item{stData}{a spot(rows)-by-gene(columns) spatial transcriptomics matrix.}

\item{markerGs}{a list with marker genes for each cell type.}

\item{noMarkerCts}{the number of cell types without marker genes.}

\item{outDir}{the path to the output directory.}

\item{seed}{the starting value.}

\item{iterations}{the number of iterations (default: 2000).}

\item{priors}{do not change the priors if you don't know what they are.}
}
\value{
a model object

a cell type proportion matrix

a cell type-specific gene expression matrix
}
\description{
Spatial transcriptomics deconvolution with the base model of SMART
}
\examples{
\dontrun{
library(SMART)
data(MPOA)
res <- SMART_base(stData=MPOA$stMat, markerGs=MPOA$markers, noMarkerCts=1,
                  outDir='SMART_results', seed=1, iterations=2000, prior=NULL)
}
}
