% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART_covariate.R
\name{SMART_covariate}
\alias{SMART_covariate}
\title{Spatial transcriptomics deconvolution with the covariate model of SMART}
\usage{
SMART_covariate(
  stData,
  markerGs,
  covarDat,
  covars,
  noMarkerCts = 1,
  outDir = "SMART_results",
  seed = 1,
  iterations = 2000,
  priors = NULL
)
}
\arguments{
\item{stData}{a spot(rows)-by-gene(columns) spatial transcriptomics matrix.}

\item{markerGs}{a list with marker genes for each cell type.}

\item{covarDat}{a data.frame containing the covariate (column) values of each spot (row).}

\item{covars}{a character vector specifying the covariates to include.}

\item{noMarkerCts}{the number of cell types without marker genes.}

\item{outDir}{the path to the output directory.}

\item{seed}{the starting value.}

\item{iterations}{the number of iterations (default: 2000).}

\item{priors}{do not change the priors if you don't know what they are.}
}
\value{
a model object
}
\description{
Spatial transcriptomics deconvolution with the covariate model of SMART
}
\examples{
\dontrun{
library(SMART)
data(MPOA)
mod <- SMART_covariate(stData=MPOA$stMat, markerGs=MPOA$markers,
                       covarDat=MPOA$covDat,covars='sex',
                       noMarkerCts=1,outDir='SMART_results', seed=1)
res <- get_est(model=mod$model, stData=MPOA$stMat,
               covarDat=MPOA$covDat, covar='sex')
}
}
