% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART_scatterheatmap.R
\name{SMART_scatterheatmap}
\alias{SMART_scatterheatmap}
\title{a scatterheatmap showing the predicted cell type proportion of a selected cell type}
\usage{
SMART_scatterheatmap(stLoc, CT_prop, celltype, cols = c("white", "red"))
}
\arguments{
\item{stLoc}{a data.frame with spatial coordinates}

\item{CT_prop}{the cell type proportions}

\item{celltype}{the cell type to plot}

\item{cols}{the color scheme for the heatmap (a character vector of two colors)}
}
\value{
a ggplot object
}
\description{
a scatterheatmap showing the predicted cell type proportion of a selected cell type
}
\examples{
\dontrun{
SMART_scatterheatmap(visium_mouse_brain$loc,res$ct_proportions, celltype = 'Oligos')
}
}
