% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART_scatterpie.R
\name{SMART_scatterpie}
\alias{SMART_scatterpie}
\title{a scatterpie plot showing the predicted cell type proportions of all cell types}
\usage{
SMART_scatterpie(stLoc, CT_prop, cols = NULL)
}
\arguments{
\item{stLoc}{a data.frame with spatial coordinates}

\item{CT_prop}{the cell type proportions}

\item{cols}{the color scheme (a character vector of multiple colors)}
}
\value{
a ggplot object
}
\description{
a scatterpie plot showing the predicted cell type proportions of all cell types
}
\examples{
\dontrun{
SMART_scatterpie(visium_mouse_brain$loc,res$ct_proportions)
}
}
