% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART.R
\name{st_train}
\alias{st_train}
\title{Deconvolution of spatial transcriptomics data with spatial information}
\usage{
st_train(
  st_counts,
  sc_results,
  spatial_coords,
  n_hvgs = 5000,
  lambda_spatial = 0.8,
  neighbor_radius = 5,
  sigma = 0.3,
  tol = 1e-08,
  maxiter = 2000,
  save_res = TRUE,
  save_mu_csv = FALSE,
  save_dir = "."
)
}
\arguments{
\item{st_counts}{a matrix indicating the raw count expression of spots (spot x gene).}

\item{sc_results}{a list indicating the output of the \code{\link{sc_train}} function.}

\item{spatial_coords}{a matrix indicating the spatial coordinates of spots (spot x 2).}

\item{n_hvgs}{an integer indicating the number of highly variable genes selected from the intersection of genes in the scRNA-seq reference and the spatial transcriptomics data. The default is 5000.}

\item{lambda_spatial}{a value indicating the weight of the spatial smoothness term. The default is 0.8.}

\item{neighbor_radius}{a value indicating the radius to define spatial neighbors. The default is 5.}

\item{sigma}{a value indicating the parameter that determine the standard deviation of the normal distribution for the platform effect. The default is 0.3.}

\item{tol}{a value indicating the convergence threshold. The default is 1e-8.}

\item{maxiter}{an integer indicating the maximum number of iterations. The default is 2000.}

\item{save_res}{a logical value indicating whether to save the results. The default is \code{TRUE}.}

\item{save_mu_csv}{a logical value indicating whether to save the cell-type-specific gene expression as csv files. The default is \code{FALSE}.}

\item{save_dir}{a character string indicating the directory where the results are saved. The default is the current directory.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{proportion}}{a matrix of the cell type proportions for spots (spot x cell type).}
  \item{\code{beta}}{a matrix of the cell type abundance for spots (spot x cell type).}
  \item{\code{alpha}}{a matrix of the combination coefficients for spots (spot x archetype).}
  \item{\code{omega}}{a matrix of the auxiliary matrix (spot x archetype).}
  \item{\code{gamma}}{a vector of the platform effect for genes.}
  \item{\code{mu}}{a list containing the cell-type-specific gene expression for spots.}
  \item{\code{Q}}{a one-hot matrix indicating the cell type corresponding to each archetype (cell type x archetype).}
  \item{\code{unique_labels}}{a vector of unique cell types included in the scRNA-seq reference.}
  \item{\code{sigma}}{a value of the parameter that determine the standard deviation of the normal distribution for the platform effect.}
  \item{\code{phi}}{a matrix of the expression of archetypes for all cell types.}
  \item{\code{phi_all}}{a matrix of the expression of archetypes with all genes for all cell types.}
  \item{\code{run_time}}{a value of the running time for deconvolution.}
  \item{\code{peak_ram}}{a value of the peak memory usage for deconvolution.}
}
}
\description{
Deconvolution of spatial transcriptomics data with spatial information
}
