% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmorph.R
\name{brainmorph}
\alias{brainmorph}
\title{Generate a brain morph plot}
\usage{
brainmorph(setup, savepaths, lineplot = TRUE, saveplot = TRUE,
  filetype = c("tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp",
  "ps", "eps", "pdf"))
}
\arguments{
\item{setup}{(required) Argument is used to get internal reference AP and z coordinates.}

\item{savepaths}{(required) Save directory path.}

\item{lineplot}{(optional, default = TRUE) Plot horizontal line at y = 1.0.}

\item{saveplot}{(optional, default=TRUE) Save the plot.
Note: saveplot will automatically save the device without leaving the device open}

\item{filetype}{(optional, default = "tif) Image type to save as. See the \emph{type} argument in the function \emph{saveplot.}
Options: "tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp","ps", "eps", "pdf"}
}
\description{
Generate a brain morph plot (normalized to 1). For use with mapping whole brain only.
Creates a plot that can be easily modified using 'par()' to query or set graphical parameters
}
