% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmorph.R
\name{brainmorph}
\alias{brainmorph}
\title{Generate a brain morph plot \strong{(W)}}
\usage{
brainmorph(
  setup,
  saveplot = TRUE,
  filetype = c("tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp", "ps", "eps",
    "pdf"),
  aspect.ratio = 0.9,
  title_size = 25,
  x_title_size = 20,
  y_title_size = 20,
  pointsize = 4,
  linesize = 1.5,
  y_lim = c(0, 1.6),
  y_by = 0.2
)
}
\arguments{
\item{setup}{(required) Setup list from \code{\link[=setup_pl]{setup_pl()}}.}

\item{saveplot}{(optional, default=TRUE) Save the plot.
Note: saveplot will automatically save the device without leaving the device open}

\item{filetype}{(optional, default = "tif") Image type to save as. See the \emph{type} argument in the function \emph{saveplot.}
Options: "tif", "tiff", "wmf", "emf", "png", "jpg", "jpeg", "bmp","ps", "eps", "pdf"}

\item{aspect.ratio}{(optional, default = 0.9) Aspect ratio of the graph.}

\item{title_size}{(optional, default = 25)}

\item{x_title_size}{(optional, default = 20)}

\item{y_title_size}{(optional, default = 20)}

\item{pointsize}{(optional, default = 4)}

\item{linesize}{(optional, default = 1.5)}

\item{y_lim}{(optional, default = c(0, 1.6)) Lower and upper limits of y-axis.}

\item{y_by}{(optional, default = 0.2) Tick mark interval of y-axis.}
}
\description{
Generate a brain morph plot (normalized to 1). For use with mapping whole brain only.
Creates a plot that can be easily modified using 'par()' to query or set graphical parameters
}
