% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_atlas.R
\name{pull_atlas}
\alias{pull_atlas}
\title{Pull up atlas plate in graphics device.}
\usage{
pull_atlas(AP, xpos = 0, x = 0, adj = c(-1, 0), cex = 1.5,
  width = NULL, save = FALSE, new_device = TRUE,
  force_redraw = FALSE)
}
\arguments{
\item{AP}{Anterior-posterior mouse atlas coordinate.}

\item{xpos}{x-position in pixels justified to the left of the main computer
screen}

\item{x}{Coordinates where the text should be written. Argument from [text()]
function.}

\item{adj}{one or two values in [0, 1] which specify the x (and optionally y)
adjustment of the labels. On most devices values outside that interval will
also work. Argument from [text()] function.}

\item{cex}{Character expension factor. Argument from [text()] function.}

\item{force_redraw}{boolean to force pull_atlas to redraw instad of getting the image}
}
\description{
Enter in an AP coordinate from the Allen Mouse Brain Atlas and
  pull up the atlas in a graphics device.
}
