% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_atlas.R
\name{pull_atlas}
\alias{pull_atlas}
\title{Pull up atlas plate in graphics device.}
\usage{
pull_atlas(AP)
}
\arguments{
\item{AP}{Anterior-posterior mouse atlas coordinate.}
}
\description{
Enter in an AP coordinate from the Allen Mouse Brain Atlas and
}
